+++ OrganList/src/main/java/es/organlist/model/dto/ProductDTO.java
@@ -14,7 +14,7 @@ import lombok.extern.jackson.Jacksonized;
 @AllArgsConstructor
 public class ProductDTO {
 
-    private Long id;
+    private Integer id;
 
     private String name;
